+++ examples/minimal/index.html
@@ -7,9 +7,9 @@
 
 <body>
     <div id='app'></div>
-    <script src="https://unpkg.com/webamp@0.0.6/built/winamp.bundle.min.js"></script>
+    <script src="https://unpkg.com/webamp@0.0.6/built/webamp.bundle.min.js"></script>
     <script>
-        const Winamp = window.Webamp;
+        const Webamp = window.Webamp;
         new Webamp({
             initialTracks: [{
                 metaData: {
